examples/test07/trippy_3d_vis/requirements.txt
pygame==2.1.2
PyOpenGL==3.1.6